refactor pipes into own file
